refactor: move screen tracking methods into Autocapture class
Extract trackScreenView and trackScreenLeave from MixpanelInstance
into a new Autocapture class, accessible via the autocapture property.
This groups autocapture methods under a dedicated namespace, signaling
that these events carry the $mp_autocapture property and appear with
an [Auto] prefix in the Mixpanel web app.
